--- src/main/java/com/webank/oracle/transaction/oraclize/TemplateOracle.java
@@ -1,140 +1,0 @@
-package com.webank.oracle.transaction.oraclize;
-
-import java.math.BigInteger;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import org.fisco.bcos.channel.client.TransactionSucCallback;
-import org.fisco.bcos.web3j.abi.FunctionReturnDecoder;
-import org.fisco.bcos.web3j.abi.TypeReference;
-import org.fisco.bcos.web3j.abi.datatypes.Function;
-import org.fisco.bcos.web3j.abi.datatypes.Type;
-import org.fisco.bcos.web3j.abi.datatypes.Utf8String;
-import org.fisco.bcos.web3j.abi.datatypes.generated.Bytes32;
-import org.fisco.bcos.web3j.crypto.Credentials;
-import org.fisco.bcos.web3j.protocol.Web3j;
-import org.fisco.bcos.web3j.protocol.core.RemoteCall;
-import org.fisco.bcos.web3j.protocol.core.methods.response.TransactionReceipt;
-import org.fisco.bcos.web3j.tuples.generated.Tuple2;
-import org.fisco.bcos.web3j.tx.Contract;
-import org.fisco.bcos.web3j.tx.TransactionManager;
-import org.fisco.bcos.web3j.tx.gas.ContractGasProvider;
-import org.fisco.bcos.web3j.tx.txdecode.TransactionDecoder;
-
-/**
- * <p>Auto generated code.
- * <p><strong>Do not modify!</strong>
- * <p>Please use the <a href="https://docs.web3j.io/command_line.html">web3j command line tools</a>,
- * or the org.fisco.bcos.web3j.codegen.SolidityFunctionWrapperGenerator in the 
- * <a href="https://github.com/web3j/web3j/tree/master/codegen">codegen module</a> to update.
- *
- * <p>Generated with web3j version null.
- */
-@SuppressWarnings("unchecked")
-public class TemplateOracle extends Contract {
-    public static String BINARY = "";
-
-    public static final String ABI = "[{\"constant\":false,\"inputs\":[{\"indexed\":false,\"name\":\"_myid\",\"type\":\"bytes32\"},{\"indexed\":false,\"name\":\"_result\",\"type\":\"string\"}],\"name\":\"__callback\",\"outputs\":[],\"payable\":false,\"stateMutability\":\"nonpayable\",\"type\":\"function\"}]";
-
-    public static final TransactionDecoder transactionDecoder = new TransactionDecoder(ABI, BINARY);
-
-    public static final String FUNC___CALLBACK = "__callback";
-
-    @Deprecated
-    protected TemplateOracle(String contractAddress, Web3j web3j, Credentials credentials, BigInteger gasPrice, BigInteger gasLimit) {
-        super(BINARY, contractAddress, web3j, credentials, gasPrice, gasLimit);
-    }
-
-    protected TemplateOracle(String contractAddress, Web3j web3j, Credentials credentials, ContractGasProvider contractGasProvider) {
-        super(BINARY, contractAddress, web3j, credentials, contractGasProvider);
-    }
-
-    @Deprecated
-    protected TemplateOracle(String contractAddress, Web3j web3j, TransactionManager transactionManager, BigInteger gasPrice, BigInteger gasLimit) {
-        super(BINARY, contractAddress, web3j, transactionManager, gasPrice, gasLimit);
-    }
-
-    protected TemplateOracle(String contractAddress, Web3j web3j, TransactionManager transactionManager, ContractGasProvider contractGasProvider) {
-        super(BINARY, contractAddress, web3j, transactionManager, contractGasProvider);
-    }
-
-    public static TransactionDecoder getTransactionDecoder() {
-        return transactionDecoder;
-    }
-
-    public RemoteCall<TransactionReceipt> __callback(byte[] _myid, String _result) {
-        final Function function = new Function(
-                FUNC___CALLBACK, 
-                Arrays.<Type>asList(new Bytes32(_myid),
-                new Utf8String(_result)),
-                Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public void __callback(byte[] _myid, String _result, TransactionSucCallback callback) {
-        final Function function = new Function(
-                FUNC___CALLBACK, 
-                Arrays.<Type>asList(new Bytes32(_myid),
-                new Utf8String(_result)),
-                Collections.<TypeReference<?>>emptyList());
-        asyncExecuteTransaction(function, callback);
-    }
-
-    public String __callbackSeq(byte[] _myid, String _result) {
-        final Function function = new Function(
-                FUNC___CALLBACK, 
-                Arrays.<Type>asList(new Bytes32(_myid),
-                new Utf8String(_result)),
-                Collections.<TypeReference<?>>emptyList());
-        return createTransactionSeq(function);
-    }
-
-    public Tuple2<byte[], String> get__callbackInput(TransactionReceipt transactionReceipt) {
-        String data = transactionReceipt.getInput().substring(10);
-        final Function function = new Function(FUNC___CALLBACK, 
-                Arrays.<Type>asList(), 
-                Arrays.<TypeReference<?>>asList(new TypeReference<Bytes32>() {}, new TypeReference<Utf8String>() {}));
-        List<Type> results = FunctionReturnDecoder.decode(data, function.getOutputParameters());;
-        return new Tuple2<byte[], String>(
-
-                (byte[]) results.get(0).getValue(), 
-                (String) results.get(1).getValue()
-                );
-    }
-
-    @Deprecated
-    public static TemplateOracle load(String contractAddress, Web3j web3j, Credentials credentials, BigInteger gasPrice, BigInteger gasLimit) {
-        return new TemplateOracle(contractAddress, web3j, credentials, gasPrice, gasLimit);
-    }
-
-    @Deprecated
-    public static TemplateOracle load(String contractAddress, Web3j web3j, TransactionManager transactionManager, BigInteger gasPrice, BigInteger gasLimit) {
-        return new TemplateOracle(contractAddress, web3j, transactionManager, gasPrice, gasLimit);
-    }
-
-    public static TemplateOracle load(String contractAddress, Web3j web3j, Credentials credentials, ContractGasProvider contractGasProvider) {
-        return new TemplateOracle(contractAddress, web3j, credentials, contractGasProvider);
-    }
-
-    public static TemplateOracle load(String contractAddress, Web3j web3j, TransactionManager transactionManager, ContractGasProvider contractGasProvider) {
-        return new TemplateOracle(contractAddress, web3j, transactionManager, contractGasProvider);
-    }
-
-    public static RemoteCall<TemplateOracle> deploy(Web3j web3j, Credentials credentials, ContractGasProvider contractGasProvider) {
-        return deployRemoteCall(TemplateOracle.class, web3j, credentials, contractGasProvider, BINARY, "");
-    }
-
-    @Deprecated
-    public static RemoteCall<TemplateOracle> deploy(Web3j web3j, Credentials credentials, BigInteger gasPrice, BigInteger gasLimit) {
-        return deployRemoteCall(TemplateOracle.class, web3j, credentials, gasPrice, gasLimit, BINARY, "");
-    }
-
-    public static RemoteCall<TemplateOracle> deploy(Web3j web3j, TransactionManager transactionManager, ContractGasProvider contractGasProvider) {
-        return deployRemoteCall(TemplateOracle.class, web3j, transactionManager, contractGasProvider, BINARY, "");
-    }
-
-    @Deprecated
-    public static RemoteCall<TemplateOracle> deploy(Web3j web3j, TransactionManager transactionManager, BigInteger gasPrice, BigInteger gasLimit) {
-        return deployRemoteCall(TemplateOracle.class, web3j, transactionManager, gasPrice, gasLimit, BINARY, "");
-    }
-}
