+++ src/test/java/com/webank/oracle/test/oracle/CommonTest.java
@@ -1,4 +1,4 @@
-package com.webank.test.oracle; /**
+package com.webank.oracle.test.oracle; /**
  * Copyright 2014-2019  the original author or authors.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
