+++ src/test/java/com/webank/oracle/test/oracle/old/OraclizeCore.java
@@ -1,4 +1,4 @@
-package com.webank.test.oracle;
+package com.webank.oracle.test.oracle.old;
 
 import java.math.BigInteger;
 import java.util.ArrayList;
